pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dateparser>=1.2.2",
    "dnspython>=2.8.0",
    "docx2pdf>=0.1.8",
    "docxtpl>=0.20.2",
    "en-core-web-sm",
    "fitz>=0.0.1.dev2",
    "flashtext>=2.7",
    "nltk>=3.9.2",
    "numpy>=2.2.6",
    "paddleocr>=3.3.2",
    "pdf2image>=1.17.0",
    "pydantic>=2.12.4",
    "pymongo>=4.15.4",
    "pymupdf>=1.26.6",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-magic>=0.4.27",
    "python-magic-bin>=0.4.14",
    "qdrant-client>=1.16.0",
    "regex>=2025.11.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "sentence-transformers>=5.1.2",
    "spacy>=3.8.11",
    "tika>=3.1.0",
    "tqdm>=4.67.1",
    "zipfile36>=0.1.3",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
